Feature/flexibleseatingplan

+++ src/app/Http/Controllers/Admin/Events/SeatingController.php
@@ -11,6 +11,7 @@ use App\Event;
 use App\EventTicket;
 use App\EventSeating;
 use App\EventSeatingPlan;
+use app\EventSeatingPlanSeat;
 use App\EventParticipant;
 use App\EventParticipantType;
 
@@ -59,7 +60,7 @@ class SeatingController extends Controller
         $rules = [
             "name"      => "required",
             "columns"   => "required|integer",
-            "rows"      => "required|integer",
+            "rows"      => "required|integer|max:26",
             'image'     => 'image',
         ];
         $messages = [
@@ -67,6 +68,7 @@ class SeatingController extends Controller
             'columns.required'  => 'Columns is required',
             'columns.integer'   => 'Columns must be a number',
             'rows.required'     => 'Rows is required',
+            'rows.max'          => 'Max. 26 Rows are allowed',
             'rows.integer'      => 'Rows must be a number',
             'image.image'       => 'Seating image must be a image',
         ];
@@ -78,7 +80,7 @@ class SeatingController extends Controller
         $seatingPlan->name_short    = @$request->name_short;
 
         $alphabet = range('A', 'Z');
-        for ($i = 0; $i < $request->columns; $i++) {
+        for ($i = 0; $i < $request->rows; $i++) {
             $seatingHeaders[] = $alphabet[$i];
         }
         $seatingPlan->headers  = implode(',', $seatingHeaders);
@@ -116,7 +118,7 @@ class SeatingController extends Controller
     {
         $rules = [
             "columns"   => "integer",
-            "rows"      => "integer",
+            "rows"      => "integer|max:26",
             'image'     => 'image',
             'status'    => 'in:draft,preview,published',
             'locked'    => 'boolean',
@@ -125,6 +127,7 @@ class SeatingController extends Controller
         $messages = [
             'columns.integer'   => 'Columns must be a number',
             'rows.integer'      => 'Rows must be a number',
+            'rows.max'          => 'Max. 26 Rows are allowed',
             'image.image'       => 'Seating image must be a image',
             'status.in'         => 'Status must be draft or published',
             'locked.boolean'    => 'Locked must be a boolean',
@@ -149,7 +152,7 @@ class SeatingController extends Controller
         if (isset($request->rows) || isset($request->columns)) {
             if ($seatingPlan->columns != $request->columns || $seatingPlan->rows != $request->rows) {
                 $alphabet = range('A', 'Z');
-                for ($i = 0; $i < $request->columns; $i++) {
+                for ($i = 0; $i < $request->rows; $i++) {
                     $seatingHeaders[]   = $alphabet[$i];
                 }
                 $seatingPlan->headers   = implode(',', $seatingHeaders);
@@ -205,7 +208,9 @@ class SeatingController extends Controller
      */
     public function storeSeat(Event $event, EventSeatingPlan $seatingPlan, Request $request)
     {
-        if (!in_array(substr($request->seat_number_modal, 0, 1), explode(',', $seatingPlan->headers)) ||
+
+        if (
+            !in_array(substr($request->seat_number_modal, 0, 1), explode(',', $seatingPlan->headers)) ||
             substr($request->seat_number_modal, 1, 1) <= 0 ||
             substr($request->seat_number_modal, 1, 1) > $seatingPlan->rows
         ) {
@@ -213,10 +218,15 @@ class SeatingController extends Controller
             return Redirect::back();
         }
 
+        if (!isset($request->seat_status_select_modal) || trim($request->seat_status_select_modal) == '') {
+            Session::flash('alert-danger', 'A status has to be selected!');
+            return Redirect::back();
+        }
+
         if (isset($request->participant_id_modal) && trim($request->participant_id_modal) != '') {
             $clauses = ['event_participant_id' => $request->participant_id_modal];
             $previousSeat = EventSeating::where($clauses)->first();
-            if ($previousSeat != null) {
+            if ($previousSeat != null && $previousSeat->status == 'ACTIVE' && strtoupper($request->seat_status_select_modal) == 'ACTIVE') {
                 $previousSeat->delete();
             }
         }
@@ -232,7 +242,7 @@ class SeatingController extends Controller
 
         $clauses = ['event_participant_id' => $request->participant_select_modal];
         $previousSeat = EventSeating::where($clauses)->first();
-        if ($previousSeat != null) {
+        if ($previousSeat != null && $previousSeat->status == 'ACTIVE' &&  strtoupper($request->seat_status_select_modal) == 'ACTIVE') {
             $previousSeat->delete();
         }
 
@@ -243,10 +253,27 @@ class SeatingController extends Controller
             return Redirect::back();
         }
 
+        //if status is set to active, the seat has to be paired with an event participant
+        if ((!isset($request->participant_select_modal) && trim($request->participant_select_modal) == '')
+            &&  strtoupper($request->seat_status_select_modal) == 'ACTIVE'
+        ) {
+            Session::flash('alert-danger', 'Seat can not be active and not have an event participant!');
+            return Redirect::back();
+        }
+
+        //if an participant is selected for seating, status can't be inactive
+        if ((isset($request->participant_select_modal) && trim($request->participant_select_modal) != '')
+            &&  strtoupper($request->seat_status_select_modal) == 'INACTIVE'
+        ) {
+            Session::flash('alert-danger', 'Seat for event participants can not be inactive!');
+            return Redirect::back();
+        }
+
         $newSeat                         = new EventSeating();
         $newSeat->seat                   = $request->seat_number_modal;
         $newSeat->event_participant_id   = $request->participant_select_modal;
         $newSeat->event_seating_plan_id  = $seatingPlan->id;
+        $newSeat->status                 = $request->seat_status_select_modal;
 
         if (!$newSeat->save()) {
             Session::flash('alert-danger', 'Could not update Seat!');
